# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SIS Kernel Control Daemon (sisctl)",
    "description": "REST API for managing SIS kernel QEMU instances",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8871",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/v1/shell/exec": {
      "post": {
        "tags": [
          "shell"
        ],
        "summary": "Execute a shell command",
        "operationId": "shellExec",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShellCommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShellCommandResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "System busy",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Seconds to wait before retrying"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Shell not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Execution failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shell/selfcheck": {
      "post": {
        "tags": [
          "shell"
        ],
        "summary": "Run self-check tests",
        "operationId": "shellSelfcheck",
        "responses": {
          "200": {
            "description": "Self-check completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelfCheckResponse"
                }
              }
            }
          },
          "409": {
            "description": "System busy",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Shell not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Self-check failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shell/selfcheck/cancel": {
      "post": {
        "tags": [
          "shell"
        ],
        "summary": "Cancel running self-check",
        "operationId": "shellSelfcheckCancel",
        "responses": {
          "200": {
            "description": "Self-check canceled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "404": {
            "description": "No self-check running",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Cancel failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sched/priorities": {
      "post": {
        "tags": [
          "scheduling"
        ],
        "summary": "Set workload priority",
        "operationId": "schedSetPriority",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPriorityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Priority updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sched/affinity": {
      "post": {
        "tags": [
          "scheduling"
        ],
        "summary": "Set workload CPU affinity",
        "operationId": "schedSetAffinity",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAffinityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Affinity updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sched/feature": {
      "post": {
        "tags": [
          "scheduling"
        ],
        "summary": "Toggle scheduling feature",
        "operationId": "schedSetFeature",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetFeatureRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Feature toggled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchedResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ShellCommandRequest": {
        "type": "object",
        "properties": {
          "command": {
            "type": "string"
          },
          "timeout_ms": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        },
        "required": [
          "command"
        ]
      },
      "ShellCommandResponse": {
        "type": "object",
        "properties": {
          "command": {
            "type": "string"
          },
          "output": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "execution_time_ms": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "command",
          "output",
          "success",
          "execution_time_ms"
        ]
      },
      "SelfCheckResponse": {
        "type": "object",
        "properties": {
          "tests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TestResultEntry"
            }
          },
          "total": {
            "type": "integer"
          },
          "passed": {
            "type": "integer"
          },
          "failed": {
            "type": "integer"
          },
          "success": {
            "type": "boolean"
          },
          "execution_time_ms": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "tests",
          "total",
          "passed",
          "failed",
          "success",
          "execution_time_ms"
        ]
      },
      "TestResultEntry": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "passed": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "name",
          "passed",
          "timestamp"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "RFC 7807 problem+json error response",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "description": "URI reference identifying the problem type",
            "example": "/errors/query-bad-range"
          },
          "title": {
            "type": "string",
            "description": "Short, human-readable summary"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "detail": {
            "type": "string",
            "description": "Human-readable explanation"
          },
          "instance": {
            "type": "string",
            "format": "uri",
            "description": "URI reference to specific occurrence",
            "nullable": true
          },
          "error": {
            "type": "string",
            "description": "Error message (legacy field)"
          }
        },
        "required": [
          "title",
          "status",
          "detail",
          "error"
        ]
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "SetPriorityRequest": {
        "type": "object",
        "properties": {
          "pid": {
            "type": "integer"
          },
          "prio": {
            "type": "integer"
          }
        },
        "required": [
          "pid",
          "prio"
        ]
      },
      "SetAffinityRequest": {
        "type": "object",
        "properties": {
          "pid": {
            "type": "integer"
          },
          "cpuMask": {
            "type": "string"
          }
        },
        "required": [
          "pid",
          "cpuMask"
        ]
      },
      "SetFeatureRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "enable": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "enable"
        ]
      },
      "SchedResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
...
    }
  },
...
}